+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Version 3.0.4 (to be released)
 
 * Added JComboBoxSelectedItemProperty and JComboBoxSelectedIndexProperty in the Swing package
+* Moved the SimpleFormatProperty to core
+* Made the FormatTransformer and ParseTransformer use a format property to allow binding on the used Format object
 
 # Version 3.0.3